Tests/LibWeb/Text/expected/Worker/Worker-location.txt

location global object URL final segment: Worker-location.js
